feat: 다른 유저도 팔로워 목록을 볼 수 있도록 수정

+++ matzipback/src/main/java/com/matzip/matzipback/users/query/mapper/UsersFollowMapper.java
@@ -18,7 +18,7 @@ public interface UsersFollowMapper {
 
     List<FollowingUsersDTO> searchFollowerUsersByUserSeqAndPage(
             @Param("followedUserSeq") long followedUserSeq,
-            @Param("offset") int offset,
+            @Param("offset") long offset,
             @Param("size") int size
     );
